+++ src/main/java/woowacourse/auth/config/AuthenticationPrincipalConfig.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import woowacourse.auth.application.AuthService;
-import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
+import woowacourse.auth.controller.AuthenticationPrincipalArgumentResolver;
 
 @Configuration
 public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
