+++ src/main/java/woowacourse/auth/controller/AuthController.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.ui;
+package woowacourse.auth.controller;
 
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
